+++ www/src/geom/Worker05GeometryPool.ts
